The rest of the simplified REPL (minus Eval!) routine.
	modified:   readword.c

The forth input stream parses on word (or number) boundaries.

SPACE (ascii 32 0x20) and ENTER (ascii 13 0x0d) are delimiters
in this forth, marking such boundaries (think of pressing ENTER
as a special case of pressing the SPACE BAR -- outcomes are
similar, in several ways).

Echoing the ENTER keystroke to the console isn't done quite
the same way as one might expect (the ok prompt may need to
be printed, first, for example).