Newline (and spacebar) end a word or a number.
The debug message reports on word boundaries.

Backspace does not trigger a word boundary.

A future refinement: do not penalty for too
much backspace (cannot reduce the buffer tib
to a negative count of accumulated chars in
the string it's holding - but it can hold just
zero bytes of accrued data. ;)

Default case is to continue building a single
wordname or multi-digit integer's ascii
representation (held in tib in a future
revision, seemingly ;)

On branch xKM_converser_aa-br-dd-
	modified:   readword.c